src/test/java/com/kovalchuk/integer_to_roman/SolutionTest.java
package com.kovalchuk.integer_to_roman;

import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.TestInstance;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class SolutionTest {
    private Solution solution;

    @BeforeAll
    public void init() {
        solution = new Solution();
    }

    static Stream<Arguments> testData() {
        return Stream.of(
                Arguments.of(3, "III"),
                Arguments.of(58, "LVIII"),
                Arguments.of(1994, "MCMXCIV"),
                Arguments.of(11993, "MMMMMMMMMMMCMXCIII")
        );
    }

    @ParameterizedTest
    @MethodSource("testData")
    void intToRoman(int input, String expected) {
        String actual = solution.intToRoman(input);

        assertEquals(expected, actual);
    }
}
